FIX: Code cleanup
